backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
httpx
python-dotenv
agno==1.4.4
openai
edge-tts
websocket-client
